# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service
  version: 2.0.0
  description: >
    Central identity and access management service.
    Handles member registration, authentication (email/password + social OAuth),
    legacy member migration, session management, token verification,
    organisational hierarchy, group management, and permissions.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:      { type: string }
            message:   { type: string }
            requestId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string }
                  message: { type: string }
                required: [message]
          required: [code, message, requestId, details]
      required: [error]
...
    Role:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string, enum: [superadmin, admin, operator, user, custom] }
        description: { type: string, nullable: true }
        created_at:  { type: string, format: date-time }
      required: [id, name, created_at]
    Scope:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string, description: "Format: app.resource.action" }
        description: { type: string, nullable: true }
        is_global:   { type: boolean }
        is_archived: { type: boolean }
        created_at:  { type: string, format: date-time }
      required: [id, name, is_global, is_archived, created_at]
    MemberScope:
      type: object
      properties:
        scope_id:    { type: string }
        name:        { type: string }
        description: { type: string, nullable: true }
        is_global:   { type: boolean }
        granted_by:  { type: string, nullable: true }
        granted_at:  { type: string, format: date-time }
      required: [scope_id, name, is_global, granted_at]
    RoleScope:
      type: object
      properties:
        scope_id:    { type: string }
        name:        { type: string }
        description: { type: string, nullable: true }
      required: [scope_id, name]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by the identity service.
        Obtain via /v1/auth/register, /v1/auth/login, or OAuth flows.
paths:
...
  /v1/permissions/scopes:
    get:
      summary: List all scopes
      operationId: listScopes
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: includeArchived, in: query, required: false, schema: { type: boolean } }]
      responses:
        '200':
          description: Scopes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Scope' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    post:
      summary: Create a scope
      operationId: createScope
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
                isGlobal:    { type: boolean }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Scope' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Name exists, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/scopes/{scopeId}/archive:
    patch:
      summary: Archive a scope
      operationId: archiveScope
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: scopeId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Scope' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/scopes/{scopeId}/restore:
    patch:
      summary: Restore an archived scope
      operationId: restoreScope
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: scopeId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Scope' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/roles:
    get:
      summary: List all roles
      operationId: listRoles
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Role' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/roles/{roleId}/scopes:
    get:
      summary: List scopes for a role
      operationId: getRoleScopes
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: roleId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Role scopes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/RoleScope' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/members/{memberId}/role:
    post:
      summary: Assign role to member
      operationId: assignRole
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: memberId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [superadmin, admin, operator, user] }
      responses:
        '204': { description: Role assigned }
        '400': { description: Invalid role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Member not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/members/{memberId}/scopes:
    get:
      summary: View member individual scope grants
      operationId: getMemberScopes
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: memberId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Member scopes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/MemberScope' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    post:
      summary: Grant scopes to member
      operationId: grantMemberScopes
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: memberId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scopes]
              properties:
                scopes: { type: array, items: { type: string }, minItems: 1 }
      responses:
        '204': { description: Granted }
        '400': { description: Unknown scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Member not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/permissions/members/{memberId}/scopes/{scopeId}:
    delete:
      summary: Revoke scope from member
      operationId: revokeMemberScope
      tags: [Permissions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: memberId, in: path, required: true, schema: { type: string } }
        - { name: scopeId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Revoked }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }